Add utils/formatDate, Add PollService.getAll, Add date to SeeAll and Vote Pages, Finish get Poll List, Fix removeOption in CreateEdit Page, Implement OptionDelete in CreateEdit Page

+++ api/v1/polls/PollController.php
@@ -50,8 +50,12 @@ class PollController extends Controller
   {
     $body = $this->json();
 
-    $this->service->delete(
-      $body->poll_id
-    );
+    if (isset($body)) {
+      $this->service->delete(
+        $body->poll_id
+      );
+    } else {
+      http_response_code(500);
+    }
   }
 }
